Modification de la route delete et suppression des images stockées

+++ controllers/books.js
@@ -1,3 +1,4 @@
+const fs= require("fs");
 const Book= require("../models/Book");
 
 exports.getBook= (req, res, next) => {
@@ -29,16 +30,44 @@ exports.postBook = (req, res, next) => {
 
 
 exports.putBook= (req, res,next) => {
-    Book.updateOne({_id: req.params.id}, {...req.body, _id: req.params.id})
-    .then(() => res.status(200).json({message: "Livre modifié"}))
-    .catch(error => res.status(400).json({error}));
-}
+    const bookObject = req.file ? {
+        ...JSON.parse(req.body.book),
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    } : { ...req.body };
+  
+    delete bookObject._userId;
+    Book.findOne({_id: req.params.id})
+        .then((book) => {
+            if (book.userId != req.auth.userId) {
+                res.status(401).json({ message : 'Non autorisé'});
+            } else {
+                Book.updateOne({ _id: req.params.id}, { ...bookObject, _id: req.params.id})
+                .then(() => res.status(200).json({message : 'Objet modifié!'}))
+                .catch(error => res.status(401).json({ error }));
+            }
+        })
+        .catch((error) => {
+            res.status(400).json({ error });
+        });
+ };
 
 exports.deleteBook= (req,res,next) => {
-    Book.deleteOne ({_id: req.params.id})
-    .then(() => res.status(200).json({message: "Livre supprimé"}))
-    .catch(error => res.status(400).json({error}));
-}
+    Book.findOne({_id: req.params.id})
+    .then(book => {
+        if (book.userId != req.auth.userId) {
+            res.status(401).json({message: "Non autorisé"});
+        } else {
+            const filename= book.imageUrl.split('/images/')[1];
+            fs.unlink(`images/${filename}`, () => {
+                Book.deleteOne({_id: req.params.id})
+                .then(() => res.status(200).json({message: "Objet supprimé"}))
+                .catch(error => res.status(400).json({error}));
+            })
+        }
+
+    })
+    .catch(error => res.status(400).json({error}))
+};
